#52 perf: 금액 자료형 long 타입으로 변경

+++ src/main/java/com/backoffice/upjuyanolja/domain/point/entity/Point.java
@@ -30,7 +30,7 @@ public class Point extends BaseTime {
 
     @Column(nullable = false)
     @Comment("보유 포인트")
-    private int pointBalance;
+    private long pointBalance;
 
     @Column(nullable = false)
     @Enumerated(value = EnumType.STRING)
